Fixed x64 warnings and added overloads to shader methods to take single parameters.

+++ trunk/source/direct3d10/ShaderBytecode10.cpp
@@ -41,7 +41,7 @@ namespace Direct3D10
 		this->blob = blob;
 	}
 
-	ShaderBytecode::ShaderBytecode( const void* buffer, int length )
+	ShaderBytecode::ShaderBytecode( const void* buffer, long length )
 	{
 		m_Buffer = buffer;
 		m_Length = length;
@@ -78,10 +78,10 @@ namespace Direct3D10
 			return m_Buffer;
 	}
 
-	int ShaderBytecode::Length::get()
+	long ShaderBytecode::Length::get()
 	{
 		if( m_Buffer == NULL )
-			return blob->GetBufferSize();
+			return static_cast<long>( blob->GetBufferSize() );
 		else
 			return m_Length;
 	}
